v3.2.0

+++ src/CouchDB.Driver/Types/Sizes.cs
@@ -11,18 +11,18 @@ namespace CouchDB.Driver.Types
         /// The size of the database file on disk in bytes. Views indexes are not included in the calculation.
         /// </summary>
         [JsonProperty("file")]
-        public int File { get; internal set; }
+        public long File { get; internal set; }
 
         /// <summary>
         /// The uncompressed size of database contents in bytes.
         /// </summary>
         [JsonProperty("external")]
-        public int External { get; internal set; }
+        public long External { get; internal set; }
 
         /// <summary>
         /// The size of live data inside the database, in bytes.
         /// </summary>
         [JsonProperty("active")]
-        public int Active { get; internal set; }
+        public long Active { get; internal set; }
     }
 }
